# perf: optimize Iterator