--- a/lib/fetch/formdata.js
+++ b/lib/fetch/formdata.js
@@ -2,9 +2,10 @@
 
 const { isBlobLike, toUSVString, makeIterator } = require('./util')
 const { kState } = require('./symbols')
+const { kEnumerableProperty } = require('../core/util')
 const { File: UndiciFile, FileLike, isFileLike } = require('./file')
 const { webidl } = require('./webidl')
-const { Blob, File: NativeFile } = require('node:buffer')
+const { File: NativeFile } = require('node:buffer')
 
 /** @type {globalThis['File']} */
 const File = NativeFile ?? UndiciFile
@@ -158,9 +159,10 @@ class FormData {
     webidl.brandCheck(this, FormData)
 
     return makeIterator(
-      () => this[kState].map(pair => [pair.name, pair.value]),
+      () => this[kState],
       'FormData',
-      'key+value'
+      'key+value',
+      'name', 'value'
     )
   }
 
@@ -168,9 +170,10 @@ class FormData {
     webidl.brandCheck(this, FormData)
 
     return makeIterator(
-      () => this[kState].map(pair => [pair.name, pair.value]),
+      () => this[kState],
       'FormData',
-      'key'
+      'key',
+      'name', 'value'
     )
   }
 
@@ -178,9 +181,10 @@ class FormData {
     webidl.brandCheck(this, FormData)
 
     return makeIterator(
-      () => this[kState].map(pair => [pair.name, pair.value]),
+      () => this[kState],
       'FormData',
-      'value'
+      'value',
+      'name', 'value'
     )
   }
 
@@ -200,7 +204,7 @@ class FormData {
     }
 
     for (const [key, value] of this) {
-      callbackFn.apply(thisArg, [value, key, this])
+      callbackFn.call(thisArg, value, key, this)
     }
   }
 }
@@ -208,6 +212,17 @@ class FormData {
 FormData.prototype[Symbol.iterator] = FormData.prototype.entries
 
 Object.defineProperties(FormData.prototype, {
+  append: kEnumerableProperty,
+  delete: kEnumerableProperty,
+  get: kEnumerableProperty,
+  getAll: kEnumerableProperty,
+  has: kEnumerableProperty,
+  set: kEnumerableProperty,
+  entries: kEnumerableProperty,
+  keys: kEnumerableProperty,
+  values: kEnumerableProperty,
+  forEach: kEnumerableProperty,
+  [Symbol.iterator]: { enumerable: false },
   [Symbol.toStringTag]: {
     value: 'FormData',
     configurable: true
@@ -225,7 +240,7 @@ function makeEntry (name, value, filename) {
   // 1. Set name to the result of converting name into a scalar value string.
   // "To convert a string into a scalar value string, replace any surrogates
   //  with U+FFFD."
-  // see: https://nodejs.org/dist/latest-v18.x/docs/api/buffer.html#buftostringencoding-start-end
+  // see: https://nodejs.org/dist/latest-v20.x/docs/api/buffer.html#buftostringencoding-start-end
   name = Buffer.from(name).toString('utf8')
 
   // 2. If value is a string, then set value to the result of converting