--- a/lib/fetch/util.js
+++ b/lib/fetch/util.js
@@ -793,6 +793,7 @@ function makeIterator (iterator, name, kind, keyIndex = 0, valueIndex = 1) {
       object.index = index + 1
       // 13. Return the iterator result for pair and kind.
       // https://webidl.spec.whatwg.org/#iterator-result
+      // 1. Let result be a value determined by the value of kind:
       let result
       if (kind === 'key') {
         // 1. Let idlKey be pair’s key.