--- a/lib/fetch/util.js
+++ b/lib/fetch/util.js
@@ -795,6 +795,35 @@ function makeIterator (iterator, name, kind, keyIndex = 0, valueIndex = 1) {
       // https://webidl.spec.whatwg.org/#iterator-result
       // 1. Let result be a value determined by the value of kind:
       let result
+      switch (kind) {
+        case 'key':
+          // 1. Let idlKey be pair’s key.
+          // 2. Let key be the result of converting idlKey to an
+          //    ECMAScript value.
+          // 3. result is key.
+          result = key
+          break
+        case 'value':
+          // 1. Let idlValue be pair’s value.
+          // 2. Let value be the result of converting idlValue to
+          //    an ECMAScript value.
+          // 3. result is value.
+          result = value
+          break
+        case 'key+value':
+          // 1. Let idlKey be pair’s key.
+          // 2. Let idlValue be pair’s value.
+          // 3. Let key be the result of converting idlKey to an
+          //    ECMAScript value.
+          // 4. Let value be the result of converting idlValue to
+          //    an ECMAScript value.
+          // 5. Let array be ! ArrayCreate(2).
+          // 6. Call ! CreateDataProperty(array, "0", key).
+          // 7. Call ! CreateDataProperty(array, "1", value).
+          // 8. result is array.
+          result = [key, value]
+          break
+      }
       if (kind === 'key') {
         // 1. Let idlKey be pair’s key.
         // 2. Let key be the result of converting idlKey to an