--- a/lib/fetch/util.js
+++ b/lib/fetch/util.js
@@ -824,31 +824,6 @@ function makeIterator (iterator, name, kind, keyIndex = 0, valueIndex = 1) {
           result = [key, value]
           break
       }
-      if (kind === 'key') {
-        // 1. Let idlKey be pair’s key.
-        // 2. Let key be the result of converting idlKey to an
-        //    ECMAScript value.
-        // 3. result is key.
-        result = key
-      } else if (kind === 'value') {
-        // 1. Let idlValue be pair’s value.
-        // 2. Let value be the result of converting idlValue to
-        //    an ECMAScript value.
-        // 3. result is value.
-        result = value
-      } else {
-        // 1. Let idlKey be pair’s key.
-        // 2. Let idlValue be pair’s value.
-        // 3. Let key be the result of converting idlKey to an
-        //    ECMAScript value.
-        // 4. Let value be the result of converting idlValue to
-        //    an ECMAScript value.
-        // 5. Let array be ! ArrayCreate(2).
-        // 6. Call ! CreateDataProperty(array, "0", key).
-        // 7. Call ! CreateDataProperty(array, "1", value).
-        // 8. result is array.
-        result = [key, value]
-      }
       // 2. Return CreateIterResultObject(result, false).
       return {
         value: result,