## Benchmark

Script from #2560

```
benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
• FormData forEach
------------------------------------------------- -----------------------------
current     6,520 ns/iter   (3,800 ns … 2,551 µs)  4,800 ns   20.3 µs  461.1 µs
new         4,521 ns/iter   (2,933 ns … 12.42 µs)  5,790 ns  10.59 µs  12.42 µs

summary for FormData forEach
  new
   1.44x faster than current

• FormData@@iterator
------------------------------------------------- -----------------------------
current     4,391 ns/iter   (3,652 ns … 10.46 µs)  4,332 ns  8,333 ns  10.46 µs
new         3,538 ns/iter   (2,403 ns … 9,227 ns)  4,948 ns  8,872 ns  9,227 ns

summary for FormData@@iterator
  new
   1.24x faster than current
```