@mcollina 2024-02-02
Can you include some benchmarks?