@tsctx 2024-02-02
> I think per spec we need to have a copy that we pass to the iterator so that it does not see changes that occur after the iterator has been created. @KhafraDev wdyt?

What kind of case is it?

I don't know if this is the case, but if it is, the current implementation does not cover it.
It indicates that it will be reflected even if it was added after creation.

Also, since this clone is created each time next is invoked, it is clear that the clone is virtually unnecessary.

```js
const headers = new Headers([["a", "0"], ["b", "1"]])

const iterator = headers.entries();

headers.append("c", "2");

console.log([...iterator]); // [ [ 'a', '0' ], [ 'b', '1' ], [ 'c', '2' ] ]
```